requirements.txt
psycopg2-binary~=2.9.11
Flask~=3.1.3
requests~=2.32.5
pyyaml==6.0.2
redis==5.0.1